Database and seeds complete. Database connection made. Server, database, controllers and server integrations.
