+++ db/seeds.sql
@@ -1,6 +1,6 @@
 
 
 
-INSERT INTO burgers (burger) VALUES ('Kangaroo Burger', false);
-INSERT INTO burgers (burger) VALUES ('Bison Burger', false);
-INSERT INTO burgers (burger) VALUES ('Ostrich Burger', false);
+INSERT INTO burgers (burger_name, devoured) VALUES ('Kangaroo Burger', false);
+INSERT INTO burgers (burger_name, devoured) VALUES ('Bison Burger', false);
+INSERT INTO burgers (burger_name, devoured) VALUES ('Ostrich Burger', false);
